rename message templates to messaging
